feat(init): Add tamper init
Added support for the PF tamper macro and initialization.

The tamper init service will disabled and clear any interrupts on
system startup. The tamper macro is added as an option, and if enabled
the base address of the macro can be set.
No (merged) targets currently use this option and therefore it is not
added to any defconfigs.