feat(services): add COLDREBOOT_FULL_FPGA_RESET
In order to perform an Auto Update when requested by the user, we need
a mechanism for triggering a device reset. While the system controller
offers a service that will initiate an Auto Update, it happens
immediately, without allowing the user application to shut itself down
gracefully.

By enabling COLDREBOOT_IS_DEVICE_RESET, and requesting an all harts cold
reboot, a user application running can trigger Auto Update when it is
safe to do so.
Use of this option is contingent on having a tamper macro in the design,
at an address specified by the option DEVICE_RESET_MACRO_ADDRESS.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>