fix(tamper): fix the !USE_TAMPER wdog build
If USE_TAMPER is not set, its base address will not be set by the
config option. To prevent build errors, set the base address to NULL if
USE_TAMPER is not set, which is safe to do as the users of the base
address depend on USE_TAMPER symbol to begin with.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>